glitch bug & cleanup
