+++ web-app/src/store/slices/conversationSlice.ts
@@ -1,48 +1,5 @@
-import ApiService from "@/services/api-service-new";
-import { DiscoveryType } from "@/types";
-import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
-import { setViewType } from "./appViewSlice";
-import { fetchDID } from "./didSlice";
-import { fetchIndex } from "./indexSlice";
-
-type FetchConversationPayload = {
-  cID: string;
-  api: ApiService;
-};
-
-export const fetchConversation = createAsyncThunk(
-  "conversation/fetchConversation",
-  async (
-    { cID, api }: FetchConversationPayload,
-    { dispatch, rejectWithValue },
-  ) => {
-    try {
-      const conversation = await api.getConversation(cID);
-      const source = conversation.sources[0];
-      let discoveryType = "unknown";
-
-      if (source.includes("did:")) {
-        discoveryType = DiscoveryType.DID;
-        await dispatch(fetchDID({ didID: source, api })).unwrap();
-      } else {
-        // TODO: check if this is index really
-        discoveryType = DiscoveryType.INDEX;
-        await dispatch(fetchIndex({ indexID: source, api })).unwrap();
-      }
-
-      dispatch(
-        setViewType({
-          type: "conversation",
-          discoveryType,
-        }),
-      );
-
-      return conversation;
-    } catch (err: any) {
-      return rejectWithValue(err.response.data);
-    }
-  },
-);
+import { fetchConversation } from "@/store/api";
+import { createSlice } from "@reduxjs/toolkit";
 
 const conversationSlice = createSlice({
   name: "conversation",
@@ -54,6 +11,9 @@ const conversationSlice = createSlice({
   reducers: {
     setConversation: (state, action) => {
       state.data = action.payload;
+    },
+    resetConversation: (state) => {
+      state.data = null;
     },
   },
   extraReducers: (builder) => {
@@ -72,6 +32,6 @@ const conversationSlice = createSlice({
   },
 });
 
-export const { setConversation } = conversationSlice.actions;
+export const { setConversation, resetConversation } = conversationSlice.actions;
 export const selectConversation = (state: any) => state.conversation;
 export default conversationSlice.reducer;
