+++ web-app/src/store/slices/indexSlice.ts
@@ -1,137 +1,11 @@
-import ApiService, { GetItemQueryParams } from "@/services/api-service-new";
-import { DiscoveryType } from "@/types";
-import { filterValidUrls, isStreamID, removeDuplicates } from "@/utils/helper";
-import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
-import { setViewType } from "./appViewSlice";
-import { fetchDID, updateDidIndex } from "./didSlice";
-import { Indexes } from "@/types/entity";
+import {
+  addItem,
+  fetchIndex,
+  fetchIndexItems,
+  toggleUserIndex,
+} from "@/store/api";
 
-type FetchIndexPayload = {
-  indexID: string;
-  api: ApiService;
-};
-
-type FetchIndexItemsPayload = {
-  indexID: string;
-  api: ApiService;
-};
-
-type AddItemPayload = {
-  item: string;
-  api: ApiService;
-  indexID: string;
-};
-
-type ToggleUserIndexPayload = {
-  indexID: string;
-  api: ApiService;
-  toggleType: "star" | "own";
-  value: boolean;
-};
-
-export const fetchIndex = createAsyncThunk(
-  "index/fetchIndex",
-  async (
-    { indexID, api }: FetchIndexPayload,
-    { dispatch, rejectWithValue },
-  ) => {
-    try {
-      const index = await api.getIndex(indexID, {});
-      let discoveryType = "unknown";
-
-      if (index) {
-        discoveryType = DiscoveryType.INDEX;
-
-        await dispatch(
-          fetchDID({ didID: index.controllerDID.id, api }),
-        ).unwrap();
-        await dispatch(fetchIndexItems({ indexID: index.id, api })).unwrap();
-      }
-
-      dispatch(
-        setViewType({
-          type: "default",
-          discoveryType,
-        }),
-      );
-
-      return index;
-    } catch (err: any) {
-      return rejectWithValue(err.response.data);
-    }
-  },
-);
-
-export const fetchIndexItems = createAsyncThunk(
-  "index/fetchIndexItems",
-  async ({ indexID, api }: FetchIndexItemsPayload, { rejectWithValue }) => {
-    try {
-      const itemParams: GetItemQueryParams = {};
-      const items = await api.getItems(indexID, {
-        queryParams: itemParams,
-      });
-
-      console.log("items", items);
-
-      return items;
-    } catch (err: any) {
-      return rejectWithValue(err.response.data);
-    }
-  },
-);
-
-export const addItem = createAsyncThunk(
-  "index/addItem",
-  async ({ item, api, indexID }: AddItemPayload, { rejectWithValue }) => {
-    try {
-      let itemId = item;
-      if (!isStreamID(item)) {
-        const createdLink = await api.crawlLink(item);
-        itemId = createdLink.id;
-      }
-
-      const createdItem = await api.createItem(indexID, itemId);
-
-      return createdItem;
-    } catch (err: any) {
-      return rejectWithValue(err.response.data);
-    }
-  },
-);
-
-export const toggleUserIndex = createAsyncThunk(
-  "index/toggleUserIndex",
-  async (
-    { indexID, api, toggleType, value }: ToggleUserIndexPayload,
-    { dispatch, rejectWithValue, getState },
-  ) => {
-    try {
-      if (toggleType === "star") {
-        await api.starIndex(indexID, value);
-      } else if (toggleType === "own") {
-        await api.ownIndex(indexID, value);
-      }
-
-      const state: any = getState();
-      const currentIndex = state.index.data;
-      dispatch(
-        updateDidIndex({
-          indexID,
-          updatedIndex: {
-            ...currentIndex,
-            did: {
-              ...currentIndex.did,
-              [toggleType === "star" ? "starred" : "owned"]: value,
-            },
-          },
-        }),
-      );
-      return { toggleType, [toggleType]: value };
-    } catch (err: any) {
-      return rejectWithValue(err.response.data);
-    }
-  },
-);
+import { createSlice } from "@reduxjs/toolkit";
 
 const indexSlice = createSlice({
   name: "index",
